move auth layout out of landing, and add auth bg with header and footer
