+++ web-app/src/app/(auth)/components/auth-background.tsx
@@ -1,0 +1,20 @@
+import Image from "next/image";
+
+import { KanjiLogo, ThemedLogo } from "@/components/masumi-logos";
+
+export default function AuthBackground() {
+  return (
+    <div className="relative h-full w-full">
+      <Image
+        alt="auth-bg"
+        src="/backgrounds/auth-bg.png"
+        fill
+        className="rounded-xl object-cover"
+        sizes="50vw"
+      />
+      <div className="pointer-events-none absolute right-4 bottom-4">
+        <ThemedLogo LogoComponent={KanjiLogo} />
+      </div>
+    </div>
+  );
+}
