+++ web-app/src/app/(auth)/components/form/types.ts
@@ -1,0 +1,5 @@
+export type AuthNamespace =
+  | "Auth.Pages.SignIn.Form"
+  | "Auth.Pages.ForgotPassword.Form"
+  | "Auth.Pages.SignUp.Form"
+  | "Auth.Pages.ResetPassword.Form";
