+++ web-app/src/app/(auth)/error.tsx
@@ -1,0 +1,55 @@
+"use client";
+
+import Link from "next/link";
+import { useTranslations } from "next-intl";
+import { useEffect } from "react";
+
+import { Button } from "@/components/ui/button";
+import {
+  Card,
+  CardContent,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+
+export default function Error({
+  error,
+  reset,
+}: {
+  error: Error & { digest?: string };
+  reset: () => void;
+}) {
+  const t = useTranslations("Auth.Error");
+
+  useEffect(() => {
+    // Log the error to an error reporting service
+    console.error(error);
+  }, [error]);
+
+  return (
+    <div className="container mx-auto flex min-h-[80vh] items-center justify-center px-4">
+      <Card className="w-full max-w-md">
+        <CardHeader>
+          <CardTitle>{t("title")}</CardTitle>
+        </CardHeader>
+        <CardContent className="space-y-4">
+          <p className="text-muted-foreground">{t("description")}</p>
+          {error.digest && (
+            <p className="text-muted-foreground text-xs">
+              {t("errorId", { errorId: error.digest })}
+            </p>
+          )}
+        </CardContent>
+        <CardFooter className="flex flex-col gap-4">
+          <Button onClick={reset} variant="default" className="w-full">
+            {t("tryAgain")}
+          </Button>
+          <Button asChild variant="outline" className="w-full">
+            <Link href="/login">{t("goLogin")}</Link>
+          </Button>
+        </CardFooter>
+      </Card>
+    </div>
+  );
+}
