+++ web-app/src/app/(auth)/forgot-password/data.ts
@@ -1,0 +1,26 @@
+import { z } from "zod";
+
+import { emailSchema } from "@/lib/auth/data";
+import { FormData } from "@/lib/form";
+
+const forgotPasswordFormSchema = (t?: IntlTranslation<"Library.Auth.Schema">) =>
+  z.object({
+    email: emailSchema(t),
+  });
+
+type ForgotPasswordFormSchemaType = z.infer<
+  ReturnType<typeof forgotPasswordFormSchema>
+>;
+
+const forgotPasswordFormData: FormData<
+  ForgotPasswordFormSchemaType,
+  "Auth.Pages.ForgotPassword.Form"
+> = [
+  {
+    name: "email",
+    placeholderKey: "Fields.Email.placeholder",
+  },
+];
+
+export { forgotPasswordFormData, forgotPasswordFormSchema };
+export type { ForgotPasswordFormSchemaType };
