--- web-app/src/app/(landing)/(auth)/components/form/types.ts
@@ -1,5 +1,0 @@
-export type AuthNamespace =
-  | "Landing.Auth.Pages.SignIn.Form"
-  | "Landing.Auth.Pages.ForgotPassword.Form"
-  | "Landing.Auth.Pages.SignUp.Form"
-  | "Landing.Auth.Pages.ResetPassword.Form";
