--- web-app/src/app/(landing)/(auth)/forgot-password/page.tsx
@@ -1,33 +1,0 @@
-import type { Metadata } from "next";
-import { getTranslations } from "next-intl/server";
-
-import ForgotPasswordForm from "./components/form";
-import ForgotPasswordHeader from "./components/header";
-
-export async function generateMetadata(): Promise<Metadata> {
-  const t = await getTranslations("Landing.Auth.Pages.ForgotPassword.Metadata");
-
-  return {
-    title: t("title"),
-    description: t("description"),
-  };
-}
-
-interface ForgotPasswordPageProps {
-  searchParams: Promise<{ email?: string }>;
-}
-
-export default async function ForgotPassword({
-  searchParams,
-}: ForgotPasswordPageProps) {
-  const { email } = await searchParams;
-
-  return (
-    <div className="flex flex-1 flex-col">
-      <ForgotPasswordHeader />
-      <div className="flex flex-1 flex-col gap-6 p-6 pt-0">
-        <ForgotPasswordForm initialEmail={email} />
-      </div>
-    </div>
-  );
-}
