--- web-app/src/app/(landing)/(auth)/layout.tsx
@@ -1,26 +1,0 @@
-import type { Metadata } from "next";
-import { getTranslations } from "next-intl/server";
-
-export async function generateMetadata(): Promise<Metadata> {
-  const t = await getTranslations("Landing.Auth.Metadata");
-
-  return {
-    title: {
-      default: t("Title.default"),
-      template: t("Title.template"),
-    },
-    description: t("description"),
-  };
-}
-
-export default function AuthLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <div className="container flex items-center justify-center p-20">
-      <div className="mx-auto w-full max-w-md">{children}</div>
-    </div>
-  );
-}
