--- web-app/src/app/(landing)/(auth)/reset-password/components/header.tsx
@@ -1,12 +1,0 @@
-import { useTranslations } from "next-intl";
-
-export default function ResetPasswordHeader() {
-  const t = useTranslations("Landing.Auth.Pages.ResetPassword");
-
-  return (
-    <div className="space-y-2 p-6">
-      <h1 className="text-2xl font-semibold tracking-tight">{t("title")}</h1>
-      <p className="text-muted-foreground text-sm">{t("description")}</p>
-    </div>
-  );
-}
